% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accounts_receivable}
\alias{accounts_receivable}
\title{Accounts receivable}
\format{
A tibble with 7 variables in the following order:
}
\usage{
accounts_receivable
}
\description{
A dataset containing 10,000 invoices outstanding per year-end and their
amount. Two audit values have also been provided.
}
\section{Fields}{

\describe{
\item{\code{debtor}}{Unique ID of the debtor.}

\item{\code{invoice}}{Unique invoice number.}

\item{\code{amount}}{Amount of the invoice.}

\item{\code{av1}}{Audit value 1.}

\item{\code{av2}}{Audit value 2.}

\item{\code{diff1}}{Difference between amount and audit value 1.}

\item{\code{diff2}}{difference between amount and audit value 2.}
}}

\keyword{datasets}
