% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{att_obj}
\alias{att_obj}
\title{Attribute state object.}
\usage{
att_obj(...)
}
\arguments{
\item{...}{Arguments passed to or from other methods.}
}
\value{
The \code{att_obj} is updated after running the \code{size} method,
  by assigning the \code{n} attribute.
}
\description{
Attribute sample state object. It contains information
  pertaining to the sample, and it is filled in consecutive stages either by
  the user or by a method applied to the object.
}
\details{
A sampling procedure consists of several stages. Some of the stages
  require the data from earlier stages. The state object records such inputs
  and provides them to later methods as required. The benefit is that these
  inputs do not have to be provided by the user repeatedly, but are obtained
  form the recorded value in the object.

  For now, the workflow of an attribute sampling procedure only consists of
  the sample size calculation stage. This produces a sample size, where the
  sample evaluation is simply the comparison of the number of deviations
  found \code{k} versus the critical number \code{c}, or the percentage of
  deviations found \code{k} / \code{n} versus the expected deviation range
  \code{eer}.
}
\examples{
myAttObject <- att_obj(alpha = .1,
                       popdev = 60,
                       popn = 1200,
                       c = 0)
myAttObject <- size(myAttObject)
myAttObject$n

myAttObject <- att_obj(alpha = .1,
                       popdev = 60,
                       popn = 1200,
                       ee = 12)
myAttObject <- size(myAttObject)
myAttObject$n

myAttObject <- att_obj(alpha = .1,
                       tdr = .05,
                       c = 1,
                       dist = "binom")
myAttObject <- size(myAttObject)
myAttObject$n

myAttObject <- att_obj(alpha = .1,
                       tdr = .05,
                       c = 2,
                       dist = "pois")
myAttObject <- size(myAttObject)
myAttObject$n

n_binom_eer <- size(myAttObject,
                    alpha = .1,
                    tdr = .05,
                    eer = 0.02,
                    dist = "binom")
n_binom_eer$n

myAttObject <- att_obj(alpha = .1,
                       popdev = 60,
                       popn = 1200,
                       c = 0)
myAttObject <- size(myAttObject)
myAttObject$n


}
