% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus_size_general.R
\name{cellResult}
\alias{cellResult}
\title{MUS Cell Evaluation.}
\usage{
cellResult(taintValues, popBv, n, alpha, dist)
}
\arguments{
\item{taintValues}{Taint values. Vector with taintings in descending order.}

\item{popBv}{Population book value. Total monetary value of the population.}

\item{n}{Sample size.}

\item{alpha}{Significance level.}

\item{dist}{Distribution used. Default distribution is \code{hyper} for the
hypergeometric distribution. Other options available are \code{binom} for
the binomial distribution and \code{pois} for the Poisson distribution.}
}
\value{
Upper precision limit on the total monetary misstatement in a
  population of \code{popBv - popSign} at \code{cl} confidence.
}
\description{
Performs cell evaluation, used in sample size calculation and
evaluation.
}
