% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_checks.R
\name{checkInteger}
\alias{checkInteger}
\title{Checks if variable is integer}
\usage{
checkInteger(var)
}
\arguments{
\item{var}{Variable. An input variable to a function.}
}
\description{
Some functions throw errors when an input variable doesn't have an integer
value. This then results in an R error message that is often incomprehensible
to a user.
}
\details{
This function traps when a variable has a non-integer value, either from a
single value or a vector of numeric values, and presents an error message
that better explains the cause of the function to halt.
}
