% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_checks.R
\name{checkLengths}
\alias{checkLengths}
\title{Check input vectors for equal length}
\usage{
checkLengths(vec1, vec2)
}
\arguments{
\item{vec1}{Vector. An input variable to a function.}

\item{vec2}{Vector. An input variable to a function.}
}
\description{
Some functions throw errors when two input vectors have different lengths.
This then results in an R error message that is often incomprehensible to a
user.
}
\details{
This function traps when input vectors have different lengths and presents an
error message that better explains the cause of the function to halt.
}
