% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_checks.R
\name{checkLessEqthan}
\alias{checkLessEqthan}
\title{Checks if one variable input is less than or equal to than another}
\usage{
checkLessEqthan(smaller, greater)
}
\arguments{
\item{smaller}{Variable. An input variable to a function.}

\item{greater}{Variable. An input variable to a function.}
}
\description{
Some functions throw errors when one input variable is less than or equal to
another. This then results in an R error message that is often
incomprehensible to a user.
}
\details{
This function traps when a variable is greater than another when it should
not, and presents an error message that better explains the cause of the
function to halt.
}
