% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_checks.R
\name{checkNonzero}
\alias{checkNonzero}
\title{Checks if variable input is not zero}
\usage{
checkNonzero(var)
}
\arguments{
\item{var}{Variable. An input variable to a function.}
}
\description{
Some functions throw errors when input variables are equal to zero. This
then results in an R error message that is often incomprehensible to a user.
}
\details{
This function traps zero input values, either from a single value or a
vector of numeric values, and presents an error message that better explains
the cause of the function to halt.
}
