% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_checks.R
\name{checkNumeric}
\alias{checkNumeric}
\title{Check if inputs are numeric}
\usage{
checkNumeric(var)
}
\arguments{
\item{var}{Variable An input variable or vector to a function.}
}
\description{
Some functions throw errors when input vectors are not numeric.
This then results in an R error message that is often incomprehensible to a
user.
}
\details{
This function traps when input vectors are not numeric and presents an
error message that better explains the cause of the function to halt.
}
