% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_checks.R
\name{checkOptions}
\alias{checkOptions}
\title{Checks if variable has value from list of options}
\usage{
checkOptions(var, options)
}
\arguments{
\item{var}{Variable. An input variable to a function.}

\item{options}{Options. A vector with character string options.}
}
\description{
Some functions throw errors when an input variable has a value assigned to it
that isn't part of a list of valid options. This then results in an R error
message that is often incomprehensible to a user.
}
\details{
This function traps when a variable has an invalid option and presents an
error message that better explains the cause of the function to halt.
}
