% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_checks.R
\name{checkPositive}
\alias{checkPositive}
\title{Checks if variable input is positive}
\usage{
checkPositive(var)
}
\arguments{
\item{var}{Variable. An input variable to a function.}
}
\description{
Some functions throw errors when input variables have a negative value. This
then results in an R error message that is often incomprehensible to a user.
}
\details{
This function traps negative input values, either from a single value or a
vector of numeric values, and presents an error message that better explains
the cause of the function to halt.
}
