% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_checks.R
\name{checkStrictsign}
\alias{checkStrictsign}
\title{Check if variables are either positive or negative}
\usage{
checkStrictsign(var)
}
\arguments{
\item{var}{Variable. An input variable to a function.}
}
\description{
Some functions throw errors when an input vector has a mixture of positive
and negative values assigned to it. This then results in an R error message
that is often incomprehensible to a user.
}
\details{
This function traps when a vector contains both negative and positive values
and presents an error message that better explains the cause of the function
to halt.
}
