% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_checks.R
\name{checkUnique}
\alias{checkUnique}
\title{Check if elements in a vector are unique}
\usage{
checkUnique(var)
}
\arguments{
\item{var}{Variable An input variable or vector to a function.}
}
\description{
Some functions require that a vector contains unique values. If this is not
the case, the code may not function correctly, the functionality may be
jeopardized, and/or an R error message is produced that is often
incomprehensible to a user.
}
\details{
This function traps when input vectors are not unique and presents an error
message that better explains the cause of the function to halt.
}
