% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvs_stratification.R
\name{cvs_eval}
\alias{cvs_eval}
\title{Evaluate CVS sample.}
\usage{
cvs_eval(av, sample, alloc, popBv, cl)
}
\arguments{
\item{av}{List of audit values corresponding to the selected sampling units.}

\item{sample}{The selected sample. Obtained from \code{cvs_select}.}

\item{alloc}{A data.frame containing the stratum label, estimated standard
deviations of the audit values, number of items, total book value, and
allocated sample size. Obtained from \code{cvs_size}.}

\item{popBv}{Total book value of the population.}

\item{cl}{Confidence level. A number greater than 0 and less than 1. Obtained
from \code{cvs_size}.}
}
\value{
List with results of the evaluation, including the effective number
  of degrees of freedom, a summary table with estimates, and for each of the
  four evaluation methods, a table with the sample size, number of nonzero
  differences, mean error value, and estimated regressor standard deviation,
  as well as the most likely audited amount, the most likely error, and the
  achieved precision.
}
\description{
Evaluates simple and stratified random samples with classical variables
methods, using the mean per unit, difference, ratio and regression
estimators.
}
