% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvs_stratification.R
\name{cvs_extend}
\alias{cvs_extend}
\title{Extend CVS sample.}
\usage{
cvs_extend(population, strata, sample, av, desPrec, cl, stratMeth)
}
\arguments{
\item{population}{Data frame consisting of the unique identifier, book
value and stratum label. Obtained from \code{cvs_stratify}.}

\item{strata}{Number of strata.}

\item{sample}{A data frame with unique identifier, book
value and stratum label of the selected elements in the initial sample.
Obtained from \code{cvs_select}.}

\item{av}{A list of audit values.}

\item{desPrec}{Desired precision (monetary amount).}

\item{cl}{Confidence level at which to reach the desired precision.}

\item{stratMeth}{Method used for stratification (\code{equal} or
\code{cumulative}).Obtained from \code{cvs_stratify}}
}
\value{
a data.frame with extended samples selected for a simple or
  stratified random sample
}
\description{
The function extends an initial simple or stratified CVS sample to achieve a
given desired precision.
}
