% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{cvs_obj}
\alias{cvs_obj}
\title{CVS state object.}
\usage{
cvs_obj(...)
}
\arguments{
\item{...}{Arguments passed to or from other methods.}
}
\value{
The \code{cvs_obj} is updated at every single stage in the sampling
  procedure. When the \code{bv} argument is assigned, the \code{popn} and
  \code{popBv} arguments are calculated. The \code{stratify} method assigns
  values to the \code{classes}, \code{strata} and \code{stratSumm} arguments.
  The \code{size} method requires input of \code{desPrec} and calculates the
  sample size \code{n} argument. The \code{select} method fills the
  \code{sample} argument. The \code{evaluate} method fills the \code{results}
  argument.
}
\description{
Classical Variables Sampling state object. It contains
  information pertaining to the sample, and it is filled in consecutive
  stages either by the user or by a method applied to the object.
}
\examples{
myCvsObject <- cvs_obj(id = inventoryData$item, bv = inventoryData$bv)
myCvsObject <- stratify(myCvsObject, strata = 3, classes = 50,
                        stratMeth = "equal")
myCvsObject <- size(myCvsObject,
                    desPrec = 200000)
myCvsObject <- select(myCvsObject,
                      seed = 20201209)
true_values <- inventoryData[match(myCvsObject$sample$item,
                                   inventoryData$item),
                             c("item", "av")]
myCvsObject <- evaluate(myCvsObject,
                        av = true_values$av)
myCvsObject$evalResults$`Effective df`
myCvsObject$evalResults$Estimates
myCvsObject$evalResults$`Most likely total audited amount mean`
myCvsObject$evalResults$`Most likely total error mean`
myCvsObject$evalResults$`Achieved precision mean`
myCvsObject$evalResults$`Difference estimation`
myCvsObject$evalResults$`Most likely total audited amount difference`
myCvsObject$evalResults$`Most likely total error difference`
myCvsObject$evalResults$`Achieved precision difference`
myCvsObject$evalResults$`Ratio estimation`
myCvsObject$evalResults$`Most likely total audited amount ratio`
myCvsObject$evalResults$`Most likely total error ratio`
myCvsObject$evalResults$`Achieved precision ratio`
myCvsObject$evalResults$`Regression estimation`
myCvsObject$evalResults$`Most likely total audited amount regression`
myCvsObject$evalResults$`Most likely total error regression`
myCvsObject$evalResults$`Achieved precision regression`
}
