% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvs_stratification.R
\name{cvs_select}
\alias{cvs_select}
\title{Sample selection.}
\usage{
cvs_select(population, nh, seed)
}
\arguments{
\item{population}{: A data frame consisting of population unique identifiers,
book values and stratum labels. Obtained from \code{cvs_stratify}.}

\item{nh}{: Numeric vector of sample sizes per stratum. Obtained from
\code{cvs_size}.}

\item{seed}{: Numeric, interpreted as an integer, between -2147483647
(\eqn{2^32 - 1}) and 2147483647 (\eqn{2^32 - 1}).}
}
\value{
A data.frame with unique identifier, book value and stratum label of
  the selected elements.
}
\description{
The function selects the sample as allocated to the strata.
}
