% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvs_stratification.R
\name{cvs_size}
\alias{cvs_size}
\title{Sample size calculation and allocation to strata.}
\usage{
cvs_size(bv, strat, strata, stratMeth, cl, desPrec, sdAv = NULL, alloc)
}
\arguments{
\item{bv}{Book value. A vector of book values, typically obtained from the
object. values and stratum labels. Obtained from \code{cvs_stratify}.}

\item{strat}{Stratum number. A vector the stratum numbers of each \code{bv}.
Obtained from \code{stratify}.}

\item{strata}{Number of strata used. Obtained from \code{stratify}.}

\item{stratMeth}{Stratification method: \code{equal} or \code{cumulative}.
Obtained from \code{cvs_stratify}.}

\item{cl}{Confidence level at which to reach the desired precision. A number
between 0 and 1.}

\item{desPrec}{Desired precision (monetary amount).}

\item{sdAv}{Vector with length equal to \code{strata}. Estimated standard
deviation of the population audit values. If omitted the standard deviation
of the book values is calculated and used as an approximation.}

\item{alloc}{Allocation table of sample size.}
}
\value{
A data frame containing the stratum label, estimated standard
  deviations of the audit values, number of items, total book value, and
  allocated sample size. If the stratification method used in
  \code{cvs_stratify} is \code{equal} then proportional allocation is used to
  allocate the sample size; if stratification method is \code{cumulative}
  then Neyman allocation is used.
}
\description{
The function calculates the sample size and allocates it to the strata using
either the proportional or the optimum (Neyman) method
}
