% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvs_stratification.R
\name{cvs_stratify}
\alias{cvs_stratify}
\title{Stratification.}
\usage{
cvs_stratify(
  id,
  bv,
  popn = NULL,
  n = NULL,
  strata,
  classes = 200,
  stratMeth = "equal"
)
}
\arguments{
\item{id}{Unique id. A unique identifier for population elements.}

\item{bv}{A list of population of book values.}

\item{popn}{Number of sampling units in the sampling frame. Calculated from
the length of \code{id}.}

\item{n}{Sample size. Calculated with the \code{size} method, or entered by
the user.}

\item{strata}{Desired number of strata. A strictly positive integer less than
half of the number of population elements.}

\item{classes}{Desired number of classes for use with the cumulative method.
Default is 200.}

\item{stratMeth}{Stratification method. Default method \code{equal} chooses
equal recorded values boundaries; \code{cumulative} uses the cumulative
square root of frequencies method.}
}
\value{
A list containing the following: \describe{ \item{Stratification
  method}{\code{equal} or \code{cumulative}} \item{Strata summary}{Data frame
  consisting of results per stratum, indicating the stratum label, the
  minimum book value, the maximum book value, the number of items, and the
  cumulative sum of the square root of the frequencies}
  \item{Population}{Data frame consisting of the unique identifier, book
  value and stratum label} \item{Class summary}{Data frame of intermediate
  results per class, showing minimum and maximum book value, frequency,
  square root of frequency, cumulative sum of square roots of frequency, and
  the ultimate stratum.} }
}
\description{
Stratifies a population according to the equal recorded value boundaries or
cumulative square root of frequencies methods.
}
\details{
The function assigns a stratum number from 1 to strata to each element in the
population.
}
