% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{errorPatterns}
\alias{errorPatterns}
\title{Error patterns}
\format{
A data frame with 8 variables in the following order:
}
\usage{
errorPatterns
}
\description{
A dataset consisting of 10,000 elements, with unique item number, book value,
and six different error seedings. All seeded errors amount to a 100,000
overstatement.
}
\section{Fields}{

\describe{
\item{\code{item}}{Item number. Unique identifier ranging from 1 to 10,000.}

\item{\code{bv}}{Book value. Book value of the item. Book values are normally
distributed, with a mean of zero, and range from -500 to 500.}

\item{\code{constdiff}}{Constant difference. Audit values of the item are such that
25\% of them have an overstatement of 40.}

\item{\code{rarediff}}{Rare difference. Like constant difference, but now only 5\%
of the items are incorrect by 200.}

\item{\code{diffslope}}{Difference and slope. Average difference between book and
audit value of 10. Audit values are fractionally off the book values, the
fraction being normally distributed.}

\item{\code{slope}}{Slope only. The audit value is 90\% of the book value, plus a
normally distributed error with mean 0 and sd equal to three times the book
value of the item. Errors are therefore heteroskedastic.}

\item{\code{missing}}{Missing. There are 914 100\% differences, representing missing
elements. All other items are correct.}

\item{\code{scatter}}{Scattered errors. No correlation between the book values and
the audit values.}
}}

\keyword{datasets}
