% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus_size_general.R
\name{findSizeCell}
\alias{findSizeCell}
\title{MUS size calculation for cell evaluation method.}
\usage{
findSizeCell(cl, popBv, pm, ee, dist)
}
\arguments{
\item{cl}{Confidence level. A number greater than 0 and less than 1.}

\item{popBv}{Population book value. Total book value of the elements in the
population.}

\item{pm}{Performance materiality. A positive integer.}

\item{ee}{Expected error. A positive integer less than \code{pm}.}

\item{dist}{Name of the distribution used. Default distribution is
\code{hyper} for the hypergeometric distribution. Other options available
are \code{binom} for the binomial distribution and \code{pois} for the
Poisson distribution.}
}
\value{
The minimum sample size required to conclude with \code{cl}
  confidence that the true monetary error does not exceed \code{pm} if the
  projected error doesn't exceed \code{ee}.
}
\description{
Calculates the sample size for a monetary unit sample.
}
