% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSaudit.R
\docType{package}
\name{FSaudit}
\alias{FSaudit-package}
\alias{FSaudit}
\title{FSaudit: A package for statistical sampling in auditing.}
\description{
The FSaudit package provides support for three categories of sampling
functions: attribute sampling, monetary unit sampling, and classical
variables sampling.
}
\section{Attribute sampling}{
 Attribute sampling functions include the
  calculation of sample sizes for fixed attribute sampling, as well as the
  calculation of upper and lower bounds.
}

\section{Monetary unit sampling}{
 Monetary unit sampling is applied by auditors
  as a hypothesis test to disprove the axistence of a material misstatement.
  Typical stages include the calculation of a sample size, elimination of
  significant items, sample selection, evaluation of the sample, and sample
  extension.
}

\section{Classical variables sampling}{
 Classical variables sampling includes
  estimation with or without auxiliary variables, where the sample may or may
  not be stratified. Methods that use auxiliary variables include regression
  estimation, difference estimation, and ratio estimation. Typical stages
  include stratification, calculation of the sample size, sample size
  allocation, sample selection, evaluation of the sample, and extension.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://LuHoo.github.io/FSaudit}
  \item Report bugs at \url{https://github.com/LuHoo/FSaudit/issues}
}

}
\author{
\strong{Maintainer}: Lucas Hoogduin \email{lucas.hoogduin@mac.com}

Authors:
\itemize{
  \item Marcel Boersma \email{boersma.marcel@gmail.com}
}

}
