% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventoryData}
\alias{inventoryData}
\title{Inventory data}
\format{
A data frame with 7 variables in the following order:
}
\usage{
inventoryData
}
\description{
A dataset containing the population ID, book values, and audit values with
different error distributions of 3,500 inventory items.
}
\section{Fields}{

\describe{
\item{\code{item}}{Unique identifier.}

\item{\code{bv}}{Book value.}

\item{\code{av}}{Audit value. Used in Audit Data Analysis case study.}

\item{\code{av_mus}}{Audit value with sparse errors. There are 116 100\%
differences, representing missing elements. All other items are correct.}

\item{\code{av_reg}}{Audit value with many errors. Differences between book and
audit value are best estimated by the regression estimator (slope and
intercept).}

\item{\code{av_dif}}{Audit value with many errors. Differences between book and
audit value are best estimated by the difference estimator (intercept only).}

\item{\code{av_rat}}{Audit value with many errors. Differences between book and
audit value are best estimated by the ratio estimator (slope only).}
}}

\keyword{datasets}
