% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_size_general.R
\name{lower}
\alias{lower}
\title{Lower bound on the number of deviations.}
\usage{
lower(k = 0, popn, n, alpha, dist = "hyper")
}
\arguments{
\item{k}{Number of deviations found in the sample. Integer less than
\code{n}.}

\item{popn}{Number of elements in the population. Positive integer.}

\item{n}{Sample size. Integer less than or equal to \code{popn}.}

\item{alpha}{Significance level.}

\item{dist}{Distribution used. Default distribution is "hyper". Other options
available are "binom" and "pois".}
}
\value{
The lower bound on the number of deviations in the population at a
  confidence level of \code{(1 - alpha)}.
}
\description{
Calculates the lower bound on the number of deviations, using the
hypergeometric, binomial, or Poisson distribution.
}
\examples{
lower(10, 1000000, 59, .05)
lower(10, 1000000, 59, .05, dist = "pois")
lower(10, 1000000, 59, .05, dist = "binom")
lower(k = 5, n = 59, alpha = .05, dist = "binom")
}
