% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_hyper.R
\name{lowerHyper}
\alias{lowerHyper}
\title{Hypergeometric lower bound.}
\usage{
lowerHyper(popn, n, k = 0, alpha)
}
\arguments{
\item{popn}{Number of elements in the population. Positive integer.}

\item{n}{Sample size. Positive integer less than or equal to \code{popn}.}

\item{k}{Number of deviations found in the sample. Positive integer less than
or equal to \code{n}.}

\item{alpha}{Significance level. A number greater than 0 and less than 1.}
}
\value{
The lower bound on the number of deviations in the population at a
  confidence level of \code{1 - alpha}.
}
\description{
Calculates the lower bound on the number of deviations using the
hypergeometric distribution, for sampling without replacement.
}
