% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus_size_general.R
\name{mus_eval}
\alias{mus_eval}
\title{MUS Evaluation.}
\usage{
mus_eval(
  cl,
  bv,
  av,
  popBv,
  dist = "hyper",
  evalMeth = "cell",
  aggregates = FALSE,
  bvTested
)
}
\arguments{
\item{cl}{Confidence level. A number greater than 0 and less than 1.}

\item{bv}{Vector with book values of the sampling units.}

\item{av}{Vector with audit values of the sampling units.}

\item{popBv}{Population book value. Total monetary value of the population.}

\item{dist}{Distribution used. Default distribution is \code{hyper} for the
hypergeometric distribution. Other options available are \code{binom} for
the binomial distribution and \code{pois} for the Poisson distribution.}

\item{evalMeth}{Name of the method used for MUS evaluation: \code{cell},
\code{stringer} or \code{pps}.}

\item{aggregates}{Boolean, indicating whether the sample is selected from
aggregate balances (TRUE) or not (FALSE, default).}

\item{bvTested}{Vector with book values of the detail items, if sample is
selected from aggregate balances.}
}
\value{
Upper precision limit on the total monetary misstatement in a
  population of \code{popBv - popSign} at \code{cl} confidence.
}
\description{
Evaluate a monetary unit sample by means of cell, Stringer
  bound or pps method.
}
