% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus_size_general.R
\name{mus_extend_n}
\alias{mus_extend_n}
\title{Extend MUS sample.}
\usage{
mus_extend_n(
  id,
  bv,
  sign_items,
  initial,
  pm,
  cl = 0.95,
  ee = 0,
  dist = "hyper",
  selMeth = "fixed",
  seed
)
}
\arguments{
\item{id}{Sampling unit. Unique identifier.}

\item{bv}{Vector of population book values, from which the total book value
is computed as the rounded sum of book values for subsequent computations.}

\item{sign_items}{Vector containing a data frame with unique identifier and
book value of individually significant items, a data frame with unique
identifier and book value of non-significant items, and a sample size based
on the total book value of the non-significant items. Obtained from
\code{mus_significant}.}

\item{initial}{Data fram with unique identifier and book value of initial
sample. Obtained from \code{mus_select}.}

\item{pm}{Performance materiality.}

\item{cl}{Confidence level. A number greater than 0 and less than 1. The
default value is 0.95.}

\item{ee}{Expected error. The default value is 0.}

\item{dist}{Name of the distribution used. Default distribution is
\code{hyper} for the hypergeometric distribution. Other options available
are \code{binom} for the binomial distribution and \code{pois} for the
Poisson distribution.}

\item{selMeth}{Optional. Default is \code{fixed}. Other options available are
\code{random} \code{randomized.fixed} \code{cell} \code{sieve} and
\code{modified.sieve}.}

\item{seed}{A single value, interpreted as an integer.}
}
\value{
A list containing a data frame of the extended sample, a data frame
  with the new individually significant items, and a data frame with the non-
  significant items.
}
\description{
The function extends an initial MUS sample to respond to a
  projected misstatement that is higher than the expected error.
}
\details{
In some cases auditors may want to extend a MUS sample, particularly
  if the initial sample yield a projected misstatement larger than the
  expected error, but clearly less than performance materiality. The
  \code{mus_extend_n} function facilitates extending the sample, while taking
  care of the population elements that become individually significant as the
  sample size increases.
}
