% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus_size_general.R
\name{mus_select}
\alias{mus_select}
\title{MUS Selection.}
\usage{
mus_select(
  id,
  bv,
  n,
  aggregates = FALSE,
  selMeth = "fixed",
  seed,
  pm,
  cl = 0.95,
  ee = 0,
  dist,
  evalMeth = "cell"
)
}
\arguments{
\item{id}{Unique id of the population elements.}

\item{bv}{Vector with book values of the population elements.}

\item{n}{Sample size.}

\item{aggregates}{Population elements are aggregates. Default value is FALSE.}

\item{selMeth}{Optional. By default, it takes fixed interval
selection. Otherwise, user can define the options, such as, a. random b.
fixed c. randomized.fixed d. cell e. sieve f. modified.sieve}

\item{seed}{Single value, interpreted as an integer. By default it is 1234.}

\item{pm}{Performance materiality.}

\item{cl}{Confidence level.}

\item{ee}{Expected error.}

\item{dist}{Distribution.}

\item{evalMeth}{Evaluation method.}
}
\value{
A data frame of selected sampling units, indicating the unique
  identifier, book value, and monetary unit selected within the sampling
  unit. For the \code{modified.sieve} method, additionally the original and
  final ranks are provided.
}
\description{
Select a monetary unit sample by means of random, (randomized)
  fixed-interval, cell, or (modified) sieve selection.
}
