% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus_size_general.R
\name{mus_significant}
\alias{mus_significant}
\title{MUS Selection of individually significant items.}
\usage{
mus_significant(id, bv, n = NULL, pm, cl = 0.95, ee = 0, dist, evalMeth)
}
\arguments{
\item{id}{Sampling unit.}

\item{bv}{Vector with book values of the population elements.}

\item{n}{Sample size. Is NULL by default, but could be preassigned by the
user, in which case it needs to be passed to the function.}

\item{pm}{Performance materiality.}

\item{cl}{Confidence level. A number greater than 0 and less than 1. The
default value is 0.95}

\item{ee}{Expected error. The default value is 0.}

\item{dist}{Name of the distribution used. Default distribution is
\code{hyper} for the hypergeometric distribution. Other options available
are \code{binom} for the binomial distribution and \code{pois} for the
Poisson distribution.}

\item{evalMeth}{Evaluation method.}
}
\value{
A data frame of individually significant items and one of
  non-significant items. In addition, a revised sample size for the
  population of non-significant items.
}
\description{
Eliminates individually significant items from the sampling
  frame.
}
