% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus_size_general.R
\name{mus_size}
\alias{mus_size}
\title{MUS size calculation.}
\usage{
mus_size(cl, bv, pm, ee = 0, dist = "hyper", evalMeth = "cell")
}
\arguments{
\item{cl}{Confidence level. A number greater than 0 and less than 1.}

\item{bv}{Population book value. It could be either a vector of book values,
from which the total book value is computed as the rounded sum of book
values for subsequent computations, or the total population book value.}

\item{pm}{Performance materiality. A positive integer.}

\item{ee}{Expected error. A positive integer less than \code{pm}.}

\item{dist}{Name of the distribution used. Default distribution is
\code{hyper} for the hypergeometric distribution. Other options available
are \code{binom} for the binomial distribution and \code{pois} for the
Poisson distribution.}

\item{evalMeth}{Evaluation method used. The sample size calculation is
slightly more efficient if the \code{evalMethod} is \code{cell}.}
}
\value{
The minimum sample size required to conclude with \code{cl}
  confidence that the true monetary error does not exceed \code{pm} if the
  projected error doesn't exceed \code{ee}.
}
\description{
Calculates the sample size for a monetary unit sample.
}
