% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus_hyper.R
\name{musSizeHyper}
\alias{musSizeHyper}
\title{MUS size calculation.}
\usage{
musSizeHyper(cl, bv, pm, ee, evalMeth)
}
\arguments{
\item{cl}{Confidence level. A number greater than 0 and less than 1.}

\item{bv}{Population book value. It could be either a vector of book values,
from which the total book value is computed as the rounded sum of book
values for subsequent computations, or the total population book value.}

\item{pm}{Performance materiality.}

\item{ee}{Expected error.}

\item{evalMeth}{Evaluation method.}
}
\value{
The minimum sample size required to conclude with \code{cl}
  confidence that the true monetary error does not exceed \code{pm} if the
  projected error doesn't exceed \code{ee}. Calculations use the
  hypergeometric distribution , consistent with sampling from a finite
  population without replacement.
}
\description{
Calculates the sample size for a monetary unit sample, using
  the hypergeometric distribution with the expected error as an argument.
}
