% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus_poisson.R
\name{musSizePois}
\alias{musSizePois}
\title{MUS size calculation.}
\usage{
musSizePois(cl, bv, pm, ee, evalMeth = "cell")
}
\arguments{
\item{cl}{Confidence level. A number greater than 0 and less than 1.}

\item{bv}{Population book value. It could be either a list of book values,
from which the total book value is computed as the rounded sum of book
values for subsequent computations, or the total population book value.}

\item{pm}{Performance materiality.}

\item{ee}{Expected error.}

\item{evalMeth}{Evaluation method.}
}
\value{
The minimum sample size required to conclude with \code{cl}
  confidence that the true monetary error does not exceed \code{pm} if the
  projected error doesn't exceed \code{ee}. Calculations use the Poisson
  approximation, assuming \code{pm} approaching 0 and population size
  approaching infinity.
}
\description{
Calculates the sample size for a monetary unit sample, using
  the Poisson approximation with the expected error as an argument.
}
\examples{
musSizePois(.90, inventoryData$bv, 100000, 20000)
musSizePois(.90, inventoryData$bv, 100000, 0)
musSizePois(.90, 5000000, 100000, 0)
musSizePois(.90, 5e06, 1e05, 0)
}
