% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salaries}
\alias{salaries}
\title{Salaries of 2,222 imaginary civil servants of Winesburg.}
\format{
A data frame with 5 variables in the following order:
}
\usage{
salaries
}
\description{
A dataset containing grade, step, gross salary, employee ID, and gender of
2,222 imaginary employees of the equally imaginary town of Winesburg.
}
\section{Fields}{

\describe{
\item{\code{grade}}{Grade of the individual. There are 20 grades. Civil servants
may be promoted to a higher grade.}

\item{\code{step}}{Step within the grade. Civil servants earn an additional step
within the grade until the maximum is reached.}

\item{\code{gross}}{Gross salary amount.}

\item{\code{id}}{Unique employee ID.}

\item{\code{sex}}{Gender. 0 indicates male, 1 indicates female.}
}}

\keyword{datasets}
