% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_grammar.R
\name{select}
\alias{select}
\alias{select.musStateObject}
\alias{select.cvsStateObject}
\title{Sample selection.}
\usage{
select(state, ...)

\method{select}{musStateObject}(state, ...)

\method{select}{cvsStateObject}(state, ...)
}
\arguments{
\item{state}{State object. This could be a \code{mus_obj} or a \code{cvs_obj}
object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
The \code{select} function is a generic function for sample
  selection. The function invokes particular \emph{methods} which depend on
  the class of the first argument.

The \code{select} function selects sample items of a given size
  from the sampling frame. The sample size is either calculated with the
  \code{getSampleSize} function or user-defined.

  \code{select.musStateObject} uses a selection procedure without replacement
  that is proportional to the book values. If individually significant items
  are not removed from the sampling frame prior to selection of the sample,
  they are interpreted as aggregates, and multiple monetary units can be
  selected from them.

  \code{select.cvsStateObject} selects either a simple random sample, if the
  population is not stratified, or a stratified random sample, where sampling
  units are selected with simple random sampling in each stratum.
}
