% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant_grammar.R
\name{significant}
\alias{significant}
\alias{significant.musStateObject}
\title{Removing significant items from the sampling frame.}
\usage{
significant(state, ...)

\method{significant}{musStateObject}(state, ...)
}
\arguments{
\item{state}{State object. This typically is an \code{mus_obj} object. The
\code{significant} method is not defined for other state objects.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
The \code{significant} function removes individually significant
  items from the sampling frame for an MUS experiment.
}
\details{
Prior to selecting the sample from a sampling frame, auditors may
  want to remove individually significant items, and test these separately.
  If not removed, all sampling units in the sampling frame are interpreted as
  aggregate items, from which one or more sampling units need to be selected
  by means of the indicated monetary unit.

  Not all selection methods facilitate proper indication of a selected
  monetary unit. For example, when working with aggregate items, it is not
  advised to use the \code{sieve} or \code{modified sieve} selection methods.

  Since significant items only pertain to sampling with unequal
  probabilities, this function is only defined in the context of a
  \code{musStateObject}.
}
