% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_grammar.R
\name{size}
\alias{size}
\alias{size.musStateObject}
\alias{size.cvsStateObject}
\alias{size.attStateObject}
\title{Sample size calculation.}
\usage{
size(state, ...)

\method{size}{musStateObject}(state, ...)

\method{size}{cvsStateObject}(state, ...)

\method{size}{attStateObject}(state, ...)
}
\arguments{
\item{state}{State object. The \code{size} method is applicable to
\code{att_obj}, \code{mus_obj} and \code{cvs_obj} objects.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
The \code{size} function is a generic function for sample size
  calculation. The function invokes particular \emph{methods} which depend on
  the class of the first argument.
}
\details{
The \code{size} function calculates the sample size for a MUS, CVS
  or attribute sample.

  MUS sample sizes are calculated using the confidence level, total
  population book value, performance materiality, and the expected
  misstatement. The default distribution is the hypergeometric,
  approximations using the binomial or Poisson distribution are available as
  well.

  Attribute sample sizes are calculated using the significance level, the
  number of tolerable deviations in the population, the number of items in
  the population, and either the expected deviation rate or the critical
  number.

  Classical variables sample sizes are calculated using the confidence level,
  the required precision, and the estmated standard deviations of the audit
  values. If these are omitted, the function uses the standard deviations of
  the book values instead.
}
