% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_binomial.R
\name{sizeBinom}
\alias{sizeBinom}
\title{Sample size binomial.}
\usage{
sizeBinom(alpha, tdr, c)
}
\arguments{
\item{alpha}{Significance level. A number greater than 0 and less than 1.}

\item{tdr}{Tolerable deviation rate. A number between 0 and 1.}

\item{c}{Critical number of deviations. A positive integer.}
}
\value{
The minimum sample size required to conclude with \code{cl}
  confidence that the true deviation rate does not exceed \code{tdr} if no
  more than \code{c} deviations are found in the sample. Calculations use the
  binomial approximation.
}
\description{
Calculate the sample size for a fixed attribute sample, using the binomial
approximation, with the number of critical deviations as an argument.
}
