% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_binomial.R
\name{sizeBinomEer}
\alias{sizeBinomEer}
\title{Sample size binomial.}
\usage{
sizeBinomEer(alpha, tdr, eer)
}
\arguments{
\item{alpha}{Significance level. A number greater than 0 and less than 1.}

\item{tdr}{Tolerable deviation rate. A number greater than 0 and less than 1.}

\item{eer}{Expected error rate. A number greater than 0 and less than 1, but
less than \code{tdr}.}
}
\value{
A list containing the sample size \code{n}, the critical number
  \code{c} associated with the sample size, and the sample size \code{nprev}
  associated with a critical number \code{c - 1}. The minimum sample size
  required to conclude with \code{1 - alpha} confidence that the true
  deviation rate does not exceed \code{tdr} if no more than \code{ceiling(n *
  eer)} deviations are found in the sample. Calculations use the binomial
  approximation.
}
\description{
Calculates the sample size for a fixed attribute sample, using the binomial
approximation, with expected error rate as an argument.
}
