% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_hyper.R
\name{sizeHyper}
\alias{sizeHyper}
\title{Sample size hypergeometric.}
\usage{
sizeHyper(alpha, popdev, popn, c)
}
\arguments{
\item{alpha}{Significance level. A number greater than 0 and less than 1.}

\item{popdev}{Number of deviations in the population. A positive integer less
than \code{popn}.}

\item{popn}{Number of items in population. A positive integer.}

\item{c}{Critical number of deviations. A positive integer less than
\code{popdev}.}
}
\value{
The minimum sample size required to conclude with \code{1 - alpha}
  confidence that the true number of deviations does not exceed \code{tdr} if
  no more than \code{c} deviations are found in the sample. Calculations use
  the hypergeometric distribution.
}
\description{
Calculates the sample size for a fixed attribute sample, using the
hypergeometric distribution.
}
