% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_hyper.R
\name{sizeHyperEe}
\alias{sizeHyperEe}
\title{Calculate the sample size for a fixed attribute sample, using the
hypergeometric distribution when expected error is used as an argument.}
\usage{
sizeHyperEe(alpha, popdev, popn, ee)
}
\arguments{
\item{alpha}{Significance level. A number greater than 0 and less than 1.}

\item{popdev}{Number of deviations in the population.}

\item{popn}{Number of items in population.}

\item{ee}{Expected error.}
}
\value{
A list containing the sample size \code{n}, the critical number
  \code{c} associated with the sample size, and the sample size \code{nprev}
  associated with a critical number \code{c - 1}. The minimum sample size
  required to conclude with \code{1 - alpha} confidence that the true
  deviation rate does not exceed \code{tdr} if no more than \code{ceiling(n *
  eer)} deviations are found in the sample. Calculations use the
  hypergeometric distribution.
}
\description{
Calculate the sample size for a fixed attribute sample, using the
hypergeometric distribution when expected error is used as an argument.
}
