% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_poisson.R
\name{sizePois}
\alias{sizePois}
\title{Sample size Poisson.}
\usage{
sizePois(alpha, tdr, c)
}
\arguments{
\item{alpha}{Significance level. A number greater than 0 and less than 1.}

\item{tdr}{Tolerable deviation rate. A number greater than 0 and less than 1.}

\item{c}{Critical number of deviations. A positive integer.}
}
\value{
The minimum sample size required to conclude with \code{1 - alpha}
  confidence that the true deviation rate does not exceed \code{tdr} if no
  more than \code{c} deviations are found in the sample. Calculations use the
  Poisson approximation, assuming \code{tdr} approaching 0 and population
  size approaching infinity.
}
\description{
Calculates the sample size for a fixed attribute sample, using the Poisson
distribution, with the number of critical deviations as an argument.
}
