% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus_size_general.R
\name{split_population}
\alias{split_population}
\title{Split population.}
\usage{
split_population(id, bv)
}
\arguments{
\item{id}{Sampling unit. Unique identifier.}

\item{bv}{Vector of population book values.}
}
\value{
The function returns a summary table with the number of sampling
  units and total book value for each of the positive, negative, and
  zero-value elements in the population, as well as tibbles with the unique
  identifier and book value of the population elements for each of the
  subpopulations. A summary of book values based on the split and list out
  the data corresponding to each category.
}
\description{
Splits the population into positive, negative and zero-value
  book values.
}
\details{
Monetary unit sampling is applied to populations with book values
  that are strictly positive or negative. Auditors may therefore be
  interested in those items that are not covered by the sampling procedure.
}
\examples{
split_population(inventoryData$item, inventoryData$bv)
}
