% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_size_general.R
\name{upper}
\alias{upper}
\title{Upper bound on the number of deviations.}
\usage{
upper(k = 0, popn, n, alpha, dist = "hyper")
}
\arguments{
\item{k}{Number of deviations found in the sample. Integer less than
\code{n}.}

\item{popn}{Number of elements in the population. Positive integer.}

\item{n}{Sample size. Integer less than or equal to \code{popn}.}

\item{alpha}{Significance level.}

\item{dist}{Distribution used. Default distribution is "hyper". Other options
available are "binom" and "pois".}
}
\value{
The one-sided upper bound on the number of deviations in the
  population at a confidence level of \code{(1 - alpha)}.
}
\description{
Calculates the upper bound on the number of deviations, using the
hypergeometric, binomial, or Poisson distribution.
}
\examples{
upper(0, 1000000, 59, .05)
upper(0, 1000000, 59, .05, dist = "binom")
upper(0, 1000000, 59, .05, dist = "pois")
}
